+++ lookatni-demo-2025-07-12T21-55-52-804Z/extracted/src/example.js
@@ -1,2 +1,0 @@
-console.log('This content belongs to src/example.js');
-const demo = true;